requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.30.6
websockets==12.0
pydantic==2.9.2
httpx==0.27.2
openai>=1.0.0
python-multipart==0.0.12
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
